Add User ref to contact model and set all contacts to private. Set middleware to check the connected user with contacts with jwt token

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all Contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getAllContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 })
 
 
 //@desc Create a Contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req,res) => {
     console.log("Request Body is: ", req.body);
     const {name,email,phone} = req.body;
@@ -24,6 +24,7 @@ const createContact = asyncHandler(async (req,res) => {
             name,
             email,
             phone,
+            user_id: req.user.id
         });
         res.status(201).json(contact);
     }
@@ -31,24 +32,32 @@ const createContact = asyncHandler(async (req,res) => {
 
 //@desc Get a Contact
 //@route GET /api/contact/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not Found.");
     }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(401);
+        throw new Error("You don't have permission to get other user's contacts");
+    }
     res.status(200).json(contact);
 })
 
 //@desc Update a Contact
 //@route PUT /api/contact/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not Found.");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(401);
+        throw new Error("You don't have permission to update other user's contacts");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -60,12 +69,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete a Contact
 //@route Delete /api/contact/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact Not Found.");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(401);
+        throw new Error("You don't have permission to delete other user's contacts");
     }
     await Contact.deleteOne();
     res.status(200).json(contact);
